backend/pyproject.toml
[project]
name = "aerooffers"
version = "3.0"
requires-python = ">=3.12"
dependencies = [
    "flask==3.1.2",
    "flask-cors==6.0.1",
    "flask-headers==1.0",
    "gunicorn==23.0.0",
    "jinja2==3.1.6",
    "nltk==3.9.2",
    "price-parser==0.5.0",
    "requests==2.32.5",
    "scrapy==2.13.4",
    "twisted==25.5.0",
    "azure-cosmos==4.14.2",
    "azure-identity==1.25.1",
    "defusedxml==0.7.1",
]

[dependency-groups]
dev = [
    {include-group = "lint"},
    "pytest==9.0.1",
    "pytest-cov==7.0.0",
    "assertpy==1.1",
    "testcontainers[cosmosdb]==4.13.3",
    "ddt==1.7.2",
]
lint = [
    "ruff==0.14.6",
    "mypy==1.18.2",
    "types-requests==2.32.4.20250913",
    "types-defusedxml==0.7.0.20250822",
    "types-flask-cors==6.0.0.20250809",
    "types-assertpy==1.1.0.20250502",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["tests/*", "**/__init__.py", ]

[tool.ruff.lint]
preview = true
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "C90", # mccabe (complexity)
    "F", # pyflakes
    "I", # imports
    "N", # pep8-naming
    "PERF", # perflint
    #"PT", # flake8-pytest-style
    #"S", # flake-8 bandit (forbid plain asserts, secrets detection, etc)
    "UP", # pyupgrade (better compatibility with newer python versions)
]
ignore = [
    "E501", # `ruff format --check` takes care of this, as this rule enforcing max line length is not compatible with formatter
    "D10" # disable requirement for everything needing a description
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "S105"] # no need for docstrings and secrets checks in tests

[tool.ruff.lint.isort]
order-by-type = false

[tool.mypy]
files = ["src/aerooffers", "tests"]
exclude = ".cache"
strict_equality = true
check_untyped_defs = true
disallow_untyped_defs = true
show_error_codes = true


[[tool.mypy.overrides]]
module = ["nltk.*", "distutils.util", "flask_headers", "ddt", "testcontainers.*"]
follow_untyped_imports = false
ignore_missing_imports = true
